add activity log

+++ backend/src/main/java/com/fpl/datn/service/Product/VariantAttributeValueService.java
@@ -1,5 +1,6 @@
 package com.fpl.datn.service.Product;
 
+import com.fpl.datn.repository.ProductVariantAttributeValueRepository;
 import org.springframework.stereotype.Service;
 
 import com.fpl.datn.exception.AppException;
@@ -18,10 +19,17 @@ import lombok.extern.slf4j.Slf4j;
 @Slf4j
 public class VariantAttributeValueService {
     VariantAttributeValueRepository repo;
-
+    ProductVariantAttributeValueRepository productVariantAttributeValueRepository;
     public void delete(Integer id) {
-        VariantAttributeValue variantAttributeValue =
-                repo.findById(id).orElseThrow(() -> new AppException(ErrorCode.VARIANT_NOT_EXISTED));
-        repo.delete(variantAttributeValue);
+      boolean existProductVariantAttributeValue = productVariantAttributeValueRepository.existsByAttributeValueId(id);
+      if(existProductVariantAttributeValue) {
+          throw new AppException(ErrorCode.VARIANT_ATTRIBUTE_ALREADY_USED);
+      }
+      else {
+          VariantAttributeValue variantAttributeValue =
+                  repo.findById(id).orElseThrow(() -> new AppException(ErrorCode.VARIANT_NOT_EXISTED));
+          repo.delete(variantAttributeValue);
+      }
+
     }
 }
